requirements.txt
ftfy
regex
tqdm
torch
torchvision
git+https://github.com/openai/CLIP.git
opencv-python
pyyaml
numpy
Pillow
ftfy
regex
tqdm
torch
torchvision
git+https://github.com/openai/CLIP.git
opencv-python
pyyaml
numpy
Pillow
flask
ultralytics